JSC_ENABLED parameter (android)
