+++ .env.jenkins
@@ -5,3 +5,4 @@ STUB_STATUS_GO=0
 ETHEREUM_DEV_CLUSTER=1
 MAINNET_NETWORKS_ENABLED=1
 LOG_LEVEL=debug
+JSC_ENABLED=1
