+++ .env.prod
@@ -5,3 +5,4 @@ STUB_STATUS_GO=0
 ETHEREUM_DEV_CLUSTER=0
 MAINNET_NETWORKS_ENABLED=0
 LOG_LEVEL=info
+JSC_ENABLED=0
