+++ modules/react-native-status/android/src/main/java/im/status/ethereum/module/StatusPackage.java
@@ -16,10 +16,12 @@ public class StatusPackage implements ReactPackage {
 
     private boolean debug;
     private boolean devCluster;
+    private boolean jscEnabled;
 
-    public StatusPackage (boolean debug, boolean devCluster) {
+    public StatusPackage (boolean debug, boolean devCluster, boolean jscEnabled) {
         this.debug = debug;
         this.devCluster = devCluster;
+        this.jscEnabled = jscEnabled;
     }
 
     @Override
@@ -27,7 +29,7 @@ public class StatusPackage implements ReactPackage {
         List<NativeModule> modules = new ArrayList<>();
         System.loadLibrary("statusgoraw");
         System.loadLibrary("statusgo");
-        modules.add(new StatusModule(reactContext, this.debug, this.devCluster));
+        modules.add(new StatusModule(reactContext, this.debug, this.devCluster, this.jscEnabled));
 
         return modules;
     }
