+++ src/status_im/utils/config.cljs
@@ -26,3 +26,4 @@
       string/lower-case
       keyword))
 
+(def jsc-enabled? (enabled? (get-config :JSC_ENABLED 0)))
